legger til statisk kalendericon * bruker ikke mer tid paa aa sette fokus etc. * is nice too have tho later
