+++ src/components/datovelger/Datovelger.less
@@ -1,0 +1,33 @@
+.datovelger {
+  &__wrapper {
+    margin-top: 1rem;
+    width: 148px;
+    height: 40px;
+    position: relative;
+    border-radius: 0.25rem;
+    background-color: #fff;
+    border: 1px solid #78706a;
+    display: grid;
+    grid-template-columns: auto auto;
+    grid-template-rows: 1fr;
+  }
+
+  .datepicker__container {
+    display: flex;
+    align-items: center;
+  }
+
+  &__input {
+    border: none;
+    font-size: medium;
+    width: 106px;
+  }
+  &__ikon {
+    width: 40px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 0 3px 3px 0;
+    border-left: solid 1px #78706a;
+  }
+}
