+++ src/index.less
@@ -13,4 +13,5 @@
 @import "./App.less";
 @import "søknad/Søknad.less";
 @import "./components/feilside/Feilside.less";
+@import "./components/datovelger/Datovelger.less";
 @import "./components/språkvelger/Språkvelger.less";
